mk_image_emd_npy.py for saving embedded images to numpy
