+++ ts/torch_handler/text_handler.py
@@ -12,7 +12,7 @@ import torch
 import torch.nn.functional as F
 from torchtext.data.utils import get_tokenizer
 from captum.attr import LayerIntegratedGradients
-from .base_handler import BaseHandler
+from .sem_pcyc_handler import BaseHandler
 from .contractions import CONTRACTION_MAP
 
 logger = logging.getLogger(__name__)
